Change return type on  executeUpdateDelete
This was a historical issue from the original Google SQLite
API.  This has been corrected to match the type returned
from the call to sqlite3_changes which represents the result.

+++ src/net/sqlcipher/database/SQLiteStatement.java
@@ -90,7 +90,7 @@ public class SQLiteStatement extends SQLiteProgram
         }
     }
 
-    public long executeUpdateDelete() {
+    public int executeUpdateDelete() {
         if (!mDatabase.isOpen()) {
             throw new IllegalStateException("database " + mDatabase.getPath() + " already closed");
         }
